reserve book

+++ Domain/Models/BookMemberM.cs
@@ -6,8 +6,8 @@ namespace Domain.Models
 {
     class BookMemberM
     {
-        public int BookId { get; set; }
-        public int MemberId { get; set; }
+        public long BookId { get; set; }
+        public long MemberId { get; set; }
         public DateTime Time { get; set; }
         public DateTime reserveTime { get; set; }
         public bool isReturned { get; set; }
